add image merge and database query checks, as well as using platform independent newline

+++ application/views/profile.php
@@ -127,6 +127,7 @@
 
         function mergeImages($src, $dest)
         {
+          $merge_success = false;
           $src_resource = imagecreatefrompng($src);
           $dest_resource = imagecreatefromjpeg($dest);
           $src_resource_width = imagesx($src_resource);
@@ -134,17 +135,23 @@
           $merged_img_name = "merged" . time() . ".jpg";
           $merged_img = UPLOADS_PATH . $merged_img_name;
 
-          if (imagecopy($dest_resource, $src_resource, 0, 0, 0, 0, $src_resource_width, $src_resource_height) )
-            imagejpeg($dest_resource, $merged_img);
+          if (imagecopy($dest_resource, $src_resource, 0, 0, 0, 0, $src_resource_width, $src_resource_height) && imagejpeg($dest_resource, $merged_img) )
+          {
+            $merge_success = true;
+            echo "Image successfully merged." . PHP_EOL;
+          }
           else
-            echo "Image merging failed.\n";
+            echo "Image merging failed." . PHP_EOL;
 
           imagedestroy($src_resource);
           imagedestroy($dest_resource);
 
           deleteFile($dest);
 
-          return ($merged_img_name);
+          if ($merge_success)
+            return ($merged_img_name);
+          else
+            return (false);
         }
 
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
@@ -181,28 +188,43 @@
             if ($uploaded)
             {
               $merged_img = mergeImages($effects_img, $file);
-              echo "$merged_img\n";
-              deleteFile($file);
+              echo "$merged_img" . PHP_EOL;
 
-              require_once(INCLUDES_DATABASE . "connection.inc.php");
+              if ($merged_img) {
 
-              $pdo = Db::getInstance();
-              $sql = "INSERT INTO images (user_id, filename, upload_date) VALUES (?, ?, UTC_TIMESTAMP() )";
-              $stmt = $pdo->prepare($sql);
-              $user_id = 1;
-              $stmt->bindValue(1, $user_id, PDO::PARAM_INT);
-              $stmt->bindValue(2, $merged_img, PDO::PARAM_STR);
-              $stmt->execute();
-              if ($stmt->rowCount() == 1) {
-                echo "Image successfully stored.";
+                require_once(INCLUDES_DATABASE . "connection.inc.php");
+
+                $pdo = Db::getInstance();
+                $sql = "INSERT INTO images (user_id, filename, upload_date) VALUES (?, ?, UTC_TIMESTAMP() )";
+                $stmt = $pdo->prepare($sql);
+                $user_id = 1;
+                $stmt->bindValue(1, $user_id, PDO::PARAM_INT);
+                $stmt->bindValue(2, $merged_img, PDO::PARAM_STR);
+                $result = $stmt->execute();
+                if ($result) {
+
+                  if ($stmt->rowCount() == 1) {
+                    echo "Image successfully stored." . PHP_EOL;
+                  }
+                  else {
+                    echo "Image storage failed." . PHP_EOL;
+                    deleteFile($merged_img);
+                  }
+
+                }
+                else {
+                  echo "Image storage failed." . PHP_EOL;
+                  deleteFile($merged_img);
+                }
+
+                $stmt = NULL;
+                $pdo = NULL;
+
               }
               else {
-                echo "Image storage failed.";
-                deleteFile($merged_img);
+                echo "Image merge failed." . PHP_EOL;
               }
 
-              $stmt = NULL;
-              $pdo = NULL;
             }
 
           }
@@ -243,14 +265,21 @@
         $stmt = $pdo->prepare($sql);
         $user_id = 1;
         $stmt->bindValue(1, $user_id, PDO::PARAM_INT);
-        $stmt->execute();
-        while ($row = $stmt->fetch() ) {
-          echo "<div class=\"profile_gallery_img\">
-                  <img src=\"uploads/" . $row['filename'] . "\" alt=\"\" />
-                  <div>
-                    <a href=\"?controller=profile&action=deleteImage&image_id=" . $row['image_id'] . "\">Delete</a>
-                  </div>
-                </div>";
+        $result = $stmt->execute();
+        if ($result) {
+
+          while ($row = $stmt->fetch() ) {
+            echo "<div class=\"profile_gallery_img\">
+                    <img src=\"uploads/" . $row['filename'] . "\" alt=\"\" />
+                    <div>
+                      <a href=\"?controller=profile&action=deleteImage&image_id=" . $row['image_id'] . "\">Delete</a>
+                    </div>
+                  </div>";
+          }
+
+        }
+        else {
+          echo "Image gallery could not be loaded." . PHP_EOL;
         }
 
         $stmt = NULL;
